refactor(services): Extract planning into FirestorePlanningService (fixes #69)
